+++ src/components/common/styles/errors.js
@@ -1,0 +1,6 @@
+export const styleErrors = {
+    commonText: `
+      color: red;
+      font-weight: bold;
+    `
+};