+++ src/components/ui/link.js
@@ -1,13 +1,14 @@
 import { theme } from '../../theme/theme';
 import { getCommonButton } from '../../styles';
-import { LinkTypes } from '../../components/common/ui';
+import { LinkTypes, LinkVariants } from '../../components/common/ui';
 
 class ActionButton extends HTMLElement {
     constructor() {
         super();
-        this.shadow = this.attachShadow({mode: 'closed'});
+        this.shadow = this.attachShadow({mode: 'open'});
         this.label = this.getAttribute('label') || 'Go';
         this.buttonType = this.getAttribute('type') || 'action';
+        this.variant = this.getAttribute('variant') || '';
         this.addCl = '';
     }
     
@@ -27,6 +28,14 @@ class ActionButton extends HTMLElement {
         return color;
     }
 
+    setAdditional() {
+        let addStyle = '';
+        if (this.variant === LinkVariants.thin) {
+            addStyle = 'font-weight: normal';
+        }
+        return addStyle;
+    }
+
     render() {
         const styles = getCommonButton();
         this.shadow.innerHTML = `
@@ -40,6 +49,7 @@ class ActionButton extends HTMLElement {
                     background-color: ${this.setColor()};
                     ${styles.main}
                     ${styles.hover}
+                    ${this.setAdditional()}
                 }
             </style>
             <button class="action-link ${this.addCl}">