+++ src/pages/additional/account/index.js
@@ -1,4 +1,5 @@
 import './infoTip';
+import './reminder';
 import './login';
 import './logout';
 import './icon';